+++ battle/StageBasis.java
@@ -53,10 +53,10 @@ public class StageBasis extends BattleObj {
 	public int changeDivision = -1;
 	public int buttonDelay = 0;
 	public int[] selectedUnit = {-1, -1};
-	public final double boss_spawn;
+	public final float boss_spawn;
 	public final int[] shakeCoolDown = {0, 0};
 
-	public double siz;
+	public float siz;
 	public int work_lv, money, maxMoney, cannon, maxCannon, upgradeCost, max_num, pos;
 	public int frontLineup = 0;
 	public boolean lineupChanging = false;
@@ -66,7 +66,7 @@ public class StageBasis extends BattleObj {
 	public float n_inten, temp_n_inten;
 	public int[] shake;
 	public int shakeDuration;
-	public double shakeOffset;
+	public float shakeOffset;
 
 	public int respawnTime, unitRespawnTime;
 	public Background bg;
@@ -75,7 +75,7 @@ public class StageBasis extends BattleObj {
 	/**
 	 * Real groundHeight of battle
 	 */
-	public double midH = -1, battleHeight = -1;
+	public float midH = -1, battleHeight = -1;
 	private final List<AttackAb> la = new ArrayList<>();
 	private boolean lethal = false;
 	public int themeTime;
@@ -219,8 +219,8 @@ public class StageBasis extends BattleObj {
 		return dire == 1 ? ubase : ebase;
 	}
 
-	public double getEBHP() {
-		return 100.0 * ebase.health / ebase.maxH;
+	public float getEBHP() {
+		return 100f * ebase.health / ebase.maxH;
 	}
 
 	/**
@@ -229,9 +229,9 @@ public class StageBasis extends BattleObj {
 	 * <p>
 	 * excludeLastEdge : If range is d0 ~ d1, normally it's true if d0 <= x <= d1 where x is entity's position<br>If this is true, then it will become d0 <= x < d1
 	 */
-	public List<AbEntity> inRange(int touch, int dire, double d0, double d1, boolean excludeLastEdge) {
-		double start = Math.min(d0, d1);
-		double end = Math.max(d0, d1);
+	public List<AbEntity> inRange(int touch, int dire, float d0, float d1, boolean excludeLastEdge) {
+		float start = Math.min(d0, d1);
+		float end = Math.max(d0, d1);
 
 		List<AbEntity> ans = new ArrayList<>();
 
@@ -251,13 +251,13 @@ public class StageBasis extends BattleObj {
 		return ans;
 	}
 
-	public void registerBattleDimension(double midH, double battleHeight) {
+	public void registerBattleDimension(float midH, float battleHeight) {
 		this.midH = midH;
 		this.battleHeight = battleHeight;
 	}
 
 	public void notifyUnitDeath() {
-		double percentage = ebase.health * 100.0 / ebase.maxH;
+		float percentage = ebase.health * 100f / ebase.maxH;
 
 		for(int i = 0; i < est.killCounter.length; i++) {
 			SCDef.Line line = est.s.data.datas[i];
@@ -511,7 +511,7 @@ public class StageBasis extends BattleObj {
 				EEnemy e = est.allow();
 
 				if (e != null) {
-					e.added(1, e.mark >= 1 ? boss_spawn : 700.0);
+					e.added(1, e.mark >= 1 ? boss_spawn : 700f);
 
 					le.add(e);
 					le.sort(Comparator.comparingInt(en -> en.layer));
@@ -521,7 +521,7 @@ public class StageBasis extends BattleObj {
 					else if(st.minSpawn == st.maxSpawn)
 						respawnTime = st.minSpawn;
 					else {
-						respawnTime = st.minSpawn + (int) ((st.maxSpawn - st.minSpawn) * r.nextDouble());
+						respawnTime = st.minSpawn + (int) ((st.maxSpawn - st.minSpawn) * r.nextFloat());
 					}
 				}
 			}
@@ -796,11 +796,11 @@ public class StageBasis extends BattleObj {
 		}
 	}
 
-	private double getOffset() {
+	private float getOffset() {
 		if(shake == null)
 			return 0;
 
-		return (1 - 2 * ((shake[SHAKE_DURATION] - shakeDuration) % 2)) * (1.0 * (shake[SHAKE_END] - shake[SHAKE_INITIAL]) / (shake[SHAKE_DURATION] - 1) * (shake[SHAKE_DURATION] - shakeDuration) + shake[SHAKE_INITIAL]) / SHAKE_STABILIZER;
+		return (1 - 2 * ((shake[SHAKE_DURATION] - shakeDuration) % 2)) * (1f * (shake[SHAKE_END] - shake[SHAKE_INITIAL]) / (shake[SHAKE_DURATION] - 1) * (shake[SHAKE_DURATION] - shakeDuration) + shake[SHAKE_INITIAL]) / SHAKE_STABILIZER;
 	}
 
 	private void setBackground(Identifier<Background> id) {
