+++ battle/attack/AttackAb.java
@@ -28,9 +28,9 @@ public abstract class AttackAb extends BattleObj {
 
 	protected final Proc proc;
 	protected final List<AbEntity> capt = new ArrayList<>();
-	protected double sta, end;
+	protected float sta, end;
 
-	protected AttackAb(Entity attacker, AtkModelAb ent, int ATK, ArrayList<Trait> tr, int eab, Proc pro, double p0, double p1, MaskAtk matk, int layer, boolean isLongAtk, int time) {
+	protected AttackAb(Entity attacker, AtkModelAb ent, int ATK, ArrayList<Trait> tr, int eab, Proc pro, float p0, float p1, MaskAtk matk, int layer, boolean isLongAtk, int time) {
 		this.attacker = attacker;
 		dire = ent.getDire();
 		origin = this;
@@ -49,7 +49,7 @@ public abstract class AttackAb extends BattleObj {
 		SPtr = matk != null && matk.getSPtrait();
 	}
 
-	protected AttackAb(Entity attacker, AttackAb a, double STA, double END, boolean isLongAtk) {
+	protected AttackAb(Entity attacker, AttackAb a, float STA, float END, boolean isLongAtk) {
 		this.attacker = attacker;
 		dire = a.dire;
 		origin = a.origin;
